apply patch from Sean Verret description taken from his email see below
I made the changes to The ULM code to make the flea camera work.  Here
is a summary of the changes you can pass on to Ulm:

1. I added support for FORMAT1 type cameras.  This includes the
MODE_1024x768_MONO format that we are interested in.

2. I added support for SVGA_NONCOMPRESSED_1 video Format

In order to accomplish the above tasks I defined two new variables
videoFormat and videoMode in the Device1394 section of the parameters
file.  These variables are used in VideoDevice1394 to pass the correct
parameters to libdc.

videoFormat is a string.  Valid entries are:
VGA (This equates to the previous Ulm code)
SVGA1 This adds support for the flea camera

videoMode is an integre.  Valid entries are:
0 - (Format0 cameras handle the previous Ulm code)
1 - (Format1 handles 800x600 and 1024x768 cameras)